move build_bundle to base
